# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MyGram is a free photo sharing app written in Go. People can share, view, and comment photos by everyone. Anyone can create an account by registering an email address and creating a username.",
        "title": "MyGram API",
        "contact": {
            "name": "musshal",
            "email": "musthafafaishal@gmail.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/photos": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all photos with authentication user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Fetch all photos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseDataFetchedPhoto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mygram-api_photo_utils.ResponseMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/mygram-api_photo_utils.ResponseMessage"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create and store a photo with authentication user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Store a photo",
                "parameters": [
                    {
                        "description": "Add Photo",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.AddPhoto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseDataAddedPhoto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mygram-api_photo_utils.ResponseMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/mygram-api_photo_utils.ResponseMessage"
                        }
                    }
                }
            }
        },
        "/photos/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update a photo by id with authentication user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Update a photo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Photo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Photo",
                        "name": "json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utils.UpdatePhoto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseDataUpdatedPhoto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mygram-api_photo_utils.ResponseMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/mygram-api_photo_utils.ResponseMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/mygram-api_photo_utils.ResponseMessage"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a photo by id with authentication user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "photos"
                ],
                "summary": "Delete a photo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Photo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ResponseMessageDeletedPhoto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/mygram-api_photo_utils.ResponseMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/mygram-api_photo_utils.ResponseMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/mygram-api_photo_utils.ResponseMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "mygram-api_photo_utils.ResponseMessage": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string",
                    "example": "the error explained here"
                },
                "status": {
                    "type": "string",
                    "example": "fail"
                }
            }
        },
...
        "utils.AddPhoto": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string",
                    "example": "A caption"
                },
                "photo_url": {
                    "type": "string",
                    "example": "https://www.example.com/image.jpg"
                },
                "title": {
                    "type": "string",
                    "example": "A Title"
                }
            }
        },
...
        "utils.AddedPhoto": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "photo_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "utils.FetchedPhoto": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "photo_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/utils.User"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "utils.ResponseDataAddedPhoto": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/utils.AddedPhoto"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "utils.ResponseDataFetchedPhoto": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/utils.FetchedPhoto"
                    }
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "utils.ResponseDataUpdatedPhoto": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/utils.UpdatedPhoto"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "utils.ResponseMessageDeletedPhoto": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "your photo has been successfully deleted"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "utils.UpdatePhoto": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string",
                    "example": "A new caption"
                },
                "photo_url": {
                    "type": "string",
                    "example": "https://www.example.com/new-image.jpg"
                },
                "title": {
                    "type": "string",
                    "example": "A new title"
                }
            }
        },
...
        "utils.UpdatedPhoto": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "photo_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "utils.User": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "johndoe@example.com"
                },
                "id": {
                    "type": "string",
                    "example": "here is the generated user id"
                },
                "username": {
                    "type": "string",
                    "example": "johndoe"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}